Add laplace mechanism (#29)
Add the PRV for the Laplace mechanism. We also refactor the privacy random variable namespace to have one file per mechanism. A notebook is added explaining in more detail how other custom mechanisms may be added to the accountant.
