+++ .github/workflows/test-package.yml
@@ -28,7 +28,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install flake8 pytest
-        python -m pip install .
+        python -m pip install .[extra]
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
