+++ .gitignore
@@ -127,3 +127,7 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
+
+# Notebooks
+/notebooks/*.html
+/notebooks/*.pdf
