+++ prv_accountant/__init__.py
@@ -2,4 +2,5 @@
 # Licensed under the MIT License.
 
 from .accountant import Accountant, PRVAccountant  # noqa: F401
-from .privacy_random_variables import PrivacyRandomVariable, PoissonSubsampledGaussianMechanism  # noqa: F401
+from .privacy_random_variables import PrivacyRandomVariable  # noqa: F401
+from .privacy_random_variables import PoissonSubsampledGaussianMechanism, LaplaceMechanism  # noqa: F401
