+++ prv_accountant/privacy_random_variables/poisson_subsampled_gaussian_mechanism.py
@@ -1,14 +1,14 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 
-from abc import ABC, abstractmethod
 import numpy as np
 import math
-from scipy import integrate
 from numpy import exp, sqrt
 from numpy import power as pow
 from scipy import special
 from scipy.special import erfc
+
+from .abstract_privacy_random_variable import PrivacyRandomVariable
 
 M_SQRT2 = sqrt(np.longdouble(2))
 M_PI = np.pi
@@ -19,62 +19,6 @@ def log(x):
     x_is_0 = (x == 0)
     return np.where(valid, np.log(np.where(valid, x, 1)),
                     np.where(x_is_0, -np.inf, np.nan))
-
-
-class PrivacyRandomVariable(ABC):
-    def mean(self) -> float:
-        """Compute the mean of the random variable."""
-        raise NotImplementedError(f"{type(self)} has not provided an implementation for a mean computation.")
-
-    def probability(self, a, b):
-        """Compute the probability mass on the interval [a, b]."""
-        return self.cdf(b) - self.cdf(a)
-
-    def pdf(self, t):
-        """Compute the probability density function of this random variable at point t."""
-        raise NotImplementedError(f"{type(self)} has not provided an implementation for a pdf.")
-
-    @abstractmethod
-    def cdf(self, t):
-        """Compute the cumulative distribution function of this random variable at point t."""
-        pass
-
-    def rdp(self, alpha: float) -> float:
-        """Compute RDP of this mechanism of order alpha."""
-        raise NotImplementedError(f"{type(self)} has not provided an implementation for rdp.")
-
-
-class PrivacyRandomVariableTruncated:
-    def __init__(self, prv, t_min: float, t_max: float) -> None:
-        self.prv = prv
-        self.t_min = t_min
-        self.t_max = t_max
-        self.remaining_mass = self.prv.cdf(t_max) - self.prv.cdf(t_min)
-
-    def mean(self) -> float:
-        points = [self.t_min, -1e-1, -1e-2, -1e-3, -1e-4, -1e-5, 1e-5, 1e-4, 1e-3, 1e-2, 1e-1, self.t_max]
-        m = 0.0
-        for L, R in zip(points[:-1], points[1:]):
-            I, err = integrate.quad(self.cdf, L, R)
-
-            m += (
-                R*self.cdf(R) -
-                L*self.cdf(L) -
-                I
-            )
-
-        return m
-
-    def probability(self, a, b):
-        a = np.clip(a, self.t_min, self.t_max)
-        b = np.clip(b, self.t_min, self.t_max)
-        return self.prv.probability(a, b) / self.remaining_mass
-
-    def pdf(self, t):
-        return np.where(t < self.t_min, 0, np.where(t < self.t_max, self.prv.pdf(t)/self.remaining_mass, 0))
-
-    def cdf(self, t):
-        return np.where(t < self.t_min, 0, np.where(t < self.t_max, self.prv.cdf(t)/self.remaining_mass, 1))
 
 
 class PoissonSubsampledGaussianMechanism(PrivacyRandomVariable):
